requirements.txt
faiss-cpu==1.7.4
matplotlib==3.7.2
matplotlib-inline==0.1.3
numpy==1.24.4
opencv-python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
rank-bm25==0.2.2
scipy==1.10.1
segment-anything==1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
transformers==4.30.2
webcolors==1.13

